Refactor files, update colors
