+++ projects/swimlane/ngx-ui/src/lib/components/code-editor/dracula.css
@@ -92,3 +92,10 @@
   text-decoration: underline;
   color: white !important;
 }
+
+.cm-s-dracula span.cm-mustache {
+  border-radius: 5px;
+  color: #cdd3dd;
+  background-color: #455066;
+  padding: 0px 2px 2px 2px;
+}
